upd5

+++ src/controller.h
@@ -14,7 +14,7 @@ class Controller {
     std::thread* field_multi_thread_ = nullptr;
     public:
     static Controller& GetInstance();
-    void SetFieldPixel(size_t x, size_t y, uint8_t val);
+    void SetFieldPixel(int x, int y, uint8_t val);
     void StartUI();
     void Start();
     void ThreadsJoin();
